Changes to switch how commands are registered
